+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 * Keep result text/csv/json files in `/var/tmp` by default if no other path has been set.
 * Auto-disable fullscreen live stats (and print hint) when detecting GNU screen session without "altscreen on", because the terminal contents (including previously printed results) cannot be correctly restored without altscreen.
 * Updated Slurm batch job example to use `--foreground &` for compatibility with Slurm configs that don't allow background services and added `--exclusive` parameter to ensure only one elbencho service is started per node.
+* Added new option `--svcping` to show service live stats reponse time in fullscreen live stats.
 
 ### Fixes
 * Fixed invalid internal vector element access when posix locks are used with random I/O patterns and benchmark path is a file.